# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  geregetemplateai_internal_http_datatransfers_requests.ChangePasswordRequest:
    properties:
      current_password:
        maxLength: 72
        minLength: 1
        type: string
      new_password:
        maxLength: 72
        minLength: 12
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  geregetemplateai_internal_http_datatransfers_requests.RegisterRequest:
    properties:
      email:
        maxLength: 50
        type: string
      password:
        maxLength: 72
        minLength: 12
        type: string
      username:
        maxLength: 25
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  geregetemplateai_internal_http_datatransfers_requests.SendOTPRequest:
    properties:
      email:
        maxLength: 50
        type: string
    required:
    - email
    type: object
...
  geregetemplateai_internal_http_datatransfers_requests.VerifyOTPRequest:
    properties:
      code:
        type: string
      email:
        maxLength: 50
        type: string
    required:
    - code
    - email
    type: object
...
  geregetemplateai_internal_http_datatransfers_responses.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      org_id:
        type: string
      refresh_token:
        type: string
      role_id:
        type: integer
      token:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  geregetemplateai_internal_http_handlers_v1.BaseResponse:
    properties:
      data: {}
      message:
        type: string
      request_id:
        type: string
      status:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Gerege Template AI v1.0
    url: https://github.com/snykk/go-rest-boilerplate
  description: Fiber v3 + GORM (PostgreSQL) + Redis дээр суурилсан Clean Architecture
    бүхий Go backend. Нээлттэй эхийн snykk/go-rest-boilerplate (MIT, зохиогч Najib
    Fikri)-ээс үүсэлтэй; Gin -> Fiber v3 болон sqlx -> GORM руу хөрвүүлсэн.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/snykk/go-rest-boilerplate
  title: Gerege Template AI v1.0 API
  version: "1.0"
paths:
...
  /auth/password/change:
    put:
      consumes:
      - application/json
      description: Одоогийн нууц үгийг шалгаж, шинээр сольж, цуцлалтын хязгаарыг тэмдэглэдэг
        тул одооноос өмнө олгогдсон refresh токенууд татгалзагдана.
      parameters:
      - description: Current + new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_requests.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "400":
          description: Malformed JSON body
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "401":
          description: Current password incorrect
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Баталгаажуулагдсан хэрэглэгчийн нууц үгийг солих
      tags:
      - auth
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Хэрэглэгчийн бүртгэлийг идэвхгүй төлөвт үүсгэнэ. Дуудагч нь /auth/login
        амжилттай болохоос өмнө бүртгэлийг идэвхжүүлэхийн тулд /auth/send-otp + /auth/verify-otp-г
        дагах ёстой.
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_requests.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            allOf:
            - $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_responses.UserResponse'
              type: object
        "400":
          description: Malformed JSON body
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "409":
          description: Email or username already in use
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "422":
          description: Validation error (per-field detail in data.errors)
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
      summary: Шинэ хэрэглэгч бүртгэх
      tags:
      - auth
  /auth/send-otp:
    post:
      consumes:
      - application/json
      description: 6 оронтой OTP үүсгэж, TTL-тэйгээр Redis-д хадгалж, async mailer-ээр
        и-мэйлийг дараалалд оруулна. HTTP хариу нь бодит SMTP хүргэлт дээр биш, дараалалд
        оруулах үед буцдаг.
      parameters:
      - description: Email to send OTP to
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_requests.SendOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP enqueued
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "400":
          description: Account already activated
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "404":
          description: Email not registered
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "500":
          description: Failed to enqueue mail
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
      summary: Хэрэглэгчийн email рүү OTP код илгээх
      tags:
      - auth
  /auth/verify-otp:
    post:
      consumes:
      - application/json
      description: Өгөгдсөн кодыг Redis-тэй тулгаж шалгаж, амжилттай үед хэрэглэгчийн
        active flag-г true болгоно. Brute-force-оос хамгаалагдсан — OTP_MAX_ATTEMPTS
        алдаа (өгөгдмөл 5) гарсны дараа зөв кодтой ч email нь OTP TTL цонхны турш
        түгжигдэнэ.
      parameters:
      - description: Email + OTP code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/geregetemplateai_internal_http_datatransfers_requests.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Account activated
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "400":
          description: Invalid OTP code
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "403":
          description: Locked out — too many invalid attempts
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "404":
          description: Email not registered
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/geregetemplateai_internal_http_handlers_v1.BaseResponse'
      summary: OTP кодыг шалгаж бүртгэлийг идэвхжүүлэх
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: /auth/login эсвэл /auth/refresh-аас олгогдсон Bearer хандах токен
      (access token).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"